Closes #2170.

## Bug

`EquipmentHandlerResponse.active_power_w` / `standby_power_w` were still
`int | None`, while `equipment_factors.csv` and the factor-side DTOs
(`EquipmentFactorCreate`/`Update`/`Response`) moved to `float` ~4 weeks
ago.
Any data entry enriched with a fractional factor value 500s at
`to_response()` — reproduced in stage, entry id=385685
(`standby_power_w=2.368421053`).

## Fix

Two-field type change on `EquipmentHandlerResponse`:
`int | None` → `float | None`. No other DTO needed touching — the
formula
already casts via `float()`, and the frontend has no int assumption
(TS `number`, cast at the row level).

## Regression

`test_equipment_response_accepts_fractional_power_w` validates the
response
DTO against the exact stage payload. Confirmed it fails pre-fix with the
reported `int_from_float` error and passes post-fix.

## Swept for the same bug class elsewhere

This bug's shape: a data-entry response DTO field fed from
`primary_factor.get(...)` / `factor_values.get(...)` whose type has
drifted
from the factor DTO that actually owns the value.

- Imported every module's `data_entries.py` + `factors.py`
  (`buildings`, `equipment`, `external_cloud_and_ai`, `headcount`,
  `process_emissions`, `professional_travel`, `purchase`,
`research_facilities`), diffed `model_fields` base types
(Optional/required
differences excluded — legitimate) for every field name shared between a
  data-entry DTO and a factor DTO in the same module. Zero mismatches
  remain after this fix.
- Grepped every `handlers.py` for `primary_factor.get(...)` /
`factor_values.get(...)` written straight into a response dict, to find
which response fields are factor-enriched rather than user-entered (the
  precondition for this bug class). Only `buildings`
  (`heating_kwh_per_square_meter`, `cooling_kwh_per_square_meter`,
`ventilation_kwh_per_square_meter`, `lighting_kwh_per_square_meter`)
shares
this shape with equipment — already `float` on both sides, no fix
needed.

No further PRs from this sweep.

## Plan

`docs/src/implementation-plans/2170-equipment-response-power-float.md`

+++ backend/app/modules/equipment/data_entries.py
@@ -64,8 +64,8 @@ class EquipmentHandlerResponse(DataEntryResponseGen):
     standby_usage_hours_per_week: int | None = None
     note: str | None = None
     kg_co2eq: float | None = None
-    active_power_w: int | None = None
-    standby_power_w: int | None = None
+    active_power_w: float | None = None
+    standby_power_w: float | None = None
 
     is_new: bool = False
